Implemented partner fetch API, with additional support of stats
- Stats based on subject wise
- Stats based on category wise.
- Stats based on content type wise. (assessment, collection,
assessment-external, collection-external, offline-activity etc..)

+++ src/main/java/org/gooru/missioncontrol/bootstrap/verticles/AuthVerticle.java
@@ -87,7 +87,7 @@ public class AuthVerticle extends AbstractVerticle {
 
   private Future<Void> updateSessionExpiryInRedis(String sessionToken) {
     Future<Void> future = Future.future();
-    int sessionTimeout = config().getInteger("sessionTimeoutInSeconds");
+    long sessionTimeout = config().getLong("sessionTimeoutInSeconds");
     redisClient.expire(sessionToken, sessionTimeout, updateHandler -> {
       if (updateHandler.succeeded()) {
         LOGGER.debug("expiry time of session is updated");
